Рефактор дз по конвертации в цельсии

+++ src/main/java/InterfaceHomework/Task_1/CelsiusConvertTo.java
@@ -1,9 +1,10 @@
 package InterfaceHomework.Task_1;
 
 public interface CelsiusConvertTo {
+
     double celsiusTo(double celsius, String scale);
 
-    double celsiusTo(int celsius, String scale);
+    double celsiusTo(Long celsius, String scale);
 
-    long celsiusTo(long celsius, String scale);
+    double celsiusTo(Integer celsius, String scale);
 }
